pyproject.toml
[tool.poetry]
name = "baecommerce"
version = "1.0.0"
description = ""
authors = ["fazle rabbi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
graphene-django = "^3.1.5"
django-graphql-jwt = "^0.4.0"
django-filter = "^23.3"
